+++ JsonPathLINQ.sln
@@ -15,11 +15,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{827E0CD3-B72D-47B6-A68D-7590B98EB39B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClientGo.JsonPath", "src\ClientGo.JsonPath\ClientGo.JsonPath.csproj", "{788D9724-5298-4052-A187-A63EB3CB512A}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0AB3BF05-4346-4AA6-1389-037BE0695223}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClientGo.JsonPath.Tests", "tests\ClientGo.JsonPath.Tests\ClientGo.JsonPath.Tests.csproj", "{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -55,38 +51,13 @@ Global
 		{B4317BB6-F3BA-4176-9389-9BB15406D47E}.Release|x64.Build.0 = Release|Any CPU
 		{B4317BB6-F3BA-4176-9389-9BB15406D47E}.Release|x86.ActiveCfg = Release|Any CPU
 		{B4317BB6-F3BA-4176-9389-9BB15406D47E}.Release|x86.Build.0 = Release|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Debug|x64.Build.0 = Debug|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Debug|x86.Build.0 = Debug|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Release|x64.ActiveCfg = Release|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Release|x64.Build.0 = Release|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Release|x86.ActiveCfg = Release|Any CPU
-		{788D9724-5298-4052-A187-A63EB3CB512A}.Release|x86.Build.0 = Release|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Debug|x64.Build.0 = Debug|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Debug|x86.Build.0 = Debug|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Release|x64.ActiveCfg = Release|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Release|x64.Build.0 = Release|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Release|x86.ActiveCfg = Release|Any CPU
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
+		{E48A6D2C-5609-433A-B1F2-20018F36369A} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{B4317BB6-F3BA-4176-9389-9BB15406D47E} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
-		{788D9724-5298-4052-A187-A63EB3CB512A} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
-		{D78143CD-68FB-450A-BDAF-CF105DDF1FAB} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F8180129-0E80-4703-988A-1CB31BF227C2}
